refactor(llvm): move the shared module's code into base.rs
`shared/mod.rs` carried 268 lines of the compiler itself. A mod.rs
declares modules, visibility and re-exports; the code lives in the
sibling base.rs.

Pure relocation, no behaviour change.
